Move/organize some constants (#368)
* Move/organize some constants

* Fix build

* work

* work

* Fixes

* fix build
