+++ constants.ts
@@ -1,0 +1,12 @@
+
+//AsyncPool Constants
+export const MAX_CONCURRENT_REQUESTS = 8;
+export const MAX_CONCURRENT_SUBSCRIPTON_REQUESTS = 5;
+
+export namespace keytarConstants {
+    export const serviceId: string = 'vscode-docker';
+
+    export const dockerHubTokenKey: string = 'dockerhub.token';
+    export const dockerHubUserNameKey: string = 'dockerhub.username';
+    export const dockerHubPasswordKey: string = 'dockerhub.password';
+}
