+++ explorer/models/azureRegistryNodes.ts
@@ -1,13 +1,13 @@
+import * as ContainerModels from 'azure-arm-containerregistry/lib/models';
 import { SubscriptionModels } from 'azure-arm-resource';
 import * as moment from 'moment';
 import * as path from 'path';
 import * as request from 'request-promise';
 import * as vscode from 'vscode';
 import { parseError } from 'vscode-azureextensionui';
-import * as ContainerModels from '../../node_modules/azure-arm-containerregistry/lib/models';
+import { MAX_CONCURRENT_REQUESTS } from '../../constants'
 import { AzureAccount, AzureSession } from '../../typings/azure-account.api';
 import { AsyncPool } from '../../utils/asyncpool';
-import { MAX_CONCURRENT_REQUESTS } from '../../utils/constants'
 import { NodeBase } from './nodeBase';
 import { RegistryType } from './registryType';
