+++ explorer/utils/dockerHubUtils.ts
@@ -2,6 +2,7 @@ import * as keytarType from 'keytar';
 import * as opn from 'opn';
 import request = require('request-promise');
 import * as vscode from 'vscode';
+import { keytarConstants } from '../../constants';
 import { DockerHubImageNode, DockerHubOrgNode, DockerHubRepositoryNode } from '../models/dockerHubNodes';
 import { getCoreNodeModule } from './utils';
 
@@ -82,9 +83,9 @@ export async function dockerHubLogout(): Promise<void> {
 
     const keytar: typeof keytarType = getCoreNodeModule('keytar');
     if (keytar) {
-        await keytar.deletePassword('vscode-docker', 'dockerhub.token');
-        await keytar.deletePassword('vscode-docker', 'dockerhub.password');
-        await keytar.deletePassword('vscode-docker', 'dockerhub.username');
+        await keytar.deletePassword(keytarConstants.serviceId, keytarConstants.dockerHubTokenKey);
+        await keytar.deletePassword(keytarConstants.serviceId, keytarConstants.dockerHubPasswordKey);
+        await keytar.deletePassword(keytarConstants.serviceId, keytarConstants.dockerHubUserNameKey);
     }
     _token = null;
 }
