--- utils/constants.ts
@@ -1,4 +1,0 @@
-
-//AsyncPool Constants
-export const MAX_CONCURRENT_REQUESTS = 8;
-export const MAX_CONCURRENT_SUBSCRIPTON_REQUESTS = 5;
